# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Directory API",
        "description": "API documentation for the Cart API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/authors": {
            "get": {
                "tags": [
                    "Author"
                ],
                "summary": "Retrieve a list of authors",
                "description": "Retrieve a paginated list of authors with filters and sorting.",
                "operationId": "6e791cab4f6d9bd496a9d5ddd05fdb24",
                "parameters": [
                    {
                        "name": "filter[search]",
                        "in": "query",
                        "description": "Search by last_name, first_name, patronymic or biography.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort by author fields.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "-created_at",
                                "full_name",
                                "-full_name",
                                "birth_date",
                                "-birth_date",
                                "death_date",
                                "-death_date"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Include related books.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "books"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Author list retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AuthorResponse"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User does not have permissions."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {}
                ]
            },
            "post": {
                "tags": [
                    "Author"
                ],
                "summary": "Create an author",
                "description": "Create a new author.",
                "operationId": "5d6ba5699658034e1b0f9fb42d197514",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAuthorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Author created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AuthorFullResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User does not have permissions."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/authors/{author}": {
            "get": {
                "tags": [
                    "Author"
                ],
                "summary": "Retrieve author details",
                "description": "Retrieve detailed information about a specific author.",
                "operationId": "7c537557c462cca9909ec5980263514c",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Author identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "shevchenko-taras-grigorievich"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Author details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AuthorFullResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User does not have permissions."
                    },
                    "404": {
                        "description": "Author not found."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Author"
                ],
                "summary": "Remove an author",
                "description": "Remove the specified author.",
                "operationId": "d6b7c065f13e949ca9072d148ba8dbf6",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Author identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "shevchenko-taras-grigorievich"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Author deleted successfully."
                    },
                    "401": {
                        "description": "User does not have permissions."
                    },
                    "404": {
                        "description": "Author not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Author"
                ],
                "summary": "Update an author",
                "description": "Update the specified author.",
                "operationId": "36fbfcc245f31a04f120a0d1d03428ff",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Author identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "shevchenko-taras-grigorievich"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAuthorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Author updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AuthorFullResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User does not have permissions."
                    },
                    "404": {
                        "description": "Author not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateAuthorRequest": {
                "title": "Create Author Request",
                "description": "Request payload for creation a new author.",
                "required": [
                    "last_name",
                    "first_name"
                ],
                "properties": {
                    "last_name": {
                        "description": "Lastname of the author.",
                        "type": "string",
                        "example": "Shevchenko",
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "first_name": {
                        "description": "Firstname of the author.",
                        "type": "string",
                        "example": "Taras",
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "patronymic": {
                        "description": "Patronymic of the author.",
                        "type": "string",
                        "example": "Grigorievich",
                        "nullable": true,
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "slug": {
                        "description": "Slug of the author.",
                        "type": "string",
                        "pattern": "^[a-z0-9-]+$",
                        "example": "shevchenko-taras-grigorievich",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "birth_date": {
                        "description": "Birth date of the author.",
                        "type": "string",
                        "format": "date",
                        "example": "1814-03-09",
                        "nullable": true
                    },
                    "death_date": {
                        "description": "Death date of the author.",
                        "type": "string",
                        "format": "date",
                        "example": "1861-03-10",
                        "nullable": true
                    },
                    "biography": {
                        "description": "Biography of the author.",
                        "type": "string",
                        "example": "Ukrainian poet, artist, and thinker. Born into a family of serfs, he gained his freedom through a ransom. His work became the foundation of modern Ukrainian literature.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateAuthorRequest": {
                "title": "Update Author Request",
                "description": "Request payload for updating an author.",
                "required": [],
                "properties": {
                    "last_name": {
                        "description": "Lastname of the author.",
                        "type": "string",
                        "example": "Shevchenko",
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "first_name": {
                        "description": "Firstname of the author.",
                        "type": "string",
                        "example": "Taras",
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "patronymic": {
                        "description": "Patronymic of the author.",
                        "type": "string",
                        "example": "Grigorievich",
                        "nullable": true,
                        "maxLength": 255,
                        "minLength": 2
                    },
                    "slug": {
                        "description": "Slug of the author.",
                        "type": "string",
                        "pattern": "^[a-z0-9-]+$",
                        "example": "shevchenko-taras-grigorievich",
                        "maxLength": 255
                    },
                    "birth_date": {
                        "description": "Birth date of the author.",
                        "type": "string",
                        "format": "date",
                        "example": "1814-03-09",
                        "nullable": true
                    },
                    "death_date": {
                        "description": "Death date of the author.",
                        "type": "string",
                        "format": "date",
                        "example": "1861-03-10",
                        "nullable": true
                    },
                    "biography": {
                        "description": "Biography of the author.",
                        "type": "string",
                        "example": "Ukrainian poet, artist, and thinker. Born into a family of serfs, he gained his freedom through a ransom. His work became the foundation of modern Ukrainian literature.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AuthorResponse": {
                "title": "Author Response",
                "description": "Data of a specific author.",
                "required": [
                    "id",
                    "last_name",
                    "first_name",
                    "slug"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the author.",
                        "type": "integer",
                        "example": 1
                    },
                    "last_name": {
                        "description": "Lastname of the author.",
                        "type": "string",
                        "example": "Shevchenko"
                    },
                    "first_name": {
                        "description": "Firstname of the author.",
                        "type": "string",
                        "example": "Taras"
                    },
                    "patronymic": {
                        "description": "Patronymic of the author.",
                        "type": "string",
                        "example": "Grigorievich",
                        "nullable": true
                    },
                    "slug": {
                        "description": "Slug of the author.",
                        "type": "string",
                        "example": "shevchenko-taras-grigorievich"
                    },
                    "birth_date": {
                        "description": "Birth date of the author.",
                        "type": "string",
                        "format": "date",
                        "example": "1814-03-09",
                        "nullable": true
                    },
                    "death_date": {
                        "description": "Death date of the author.",
                        "type": "string",
                        "format": "date",
                        "example": "1861-03-10",
                        "nullable": true
                    },
                    "biography": {
                        "description": "Biography of the author.",
                        "type": "string",
                        "example": "Ukrainian poet, artist, and thinker. Born into a family of serfs, he gained his freedom through a ransom. His work became the foundation of modern Ukrainian literature.",
                        "nullable": true
                    },
                    "image_url": {
                        "description": "Image url of the author.",
                        "type": "string",
                        "format": "uri",
                        "example": "http://loclhost:8080/storage/authors/author1.png",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AuthorFullResponse": {
                "title": "Author Full Response",
                "description": "Data of a specific author.",
                "required": [
                    "id",
                    "last_name",
                    "first_name",
                    "slug",
                    "books"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the author.",
                        "type": "integer",
                        "example": 1
                    },
                    "last_name": {
                        "description": "Lastname of the author.",
                        "type": "string",
                        "example": "Shevchenko"
                    },
                    "first_name": {
                        "description": "Firstname of the author.",
                        "type": "string",
                        "example": "Taras"
                    },
                    "patronymic": {
                        "description": "Patronymic of the author.",
                        "type": "string",
                        "example": "Grigorievich",
                        "nullable": true
                    },
                    "slug": {
                        "description": "Slug of the author.",
                        "type": "string",
                        "example": "shevchenko-taras-grigorievich"
                    },
                    "birth_date": {
                        "description": "Birth date of the author.",
                        "type": "string",
                        "format": "date",
                        "example": "1814-03-09",
                        "nullable": true
                    },
                    "death_date": {
                        "description": "Death date of the author.",
                        "type": "string",
                        "format": "date",
                        "example": "1861-03-10",
                        "nullable": true
                    },
                    "biography": {
                        "description": "Biography of the author.",
                        "type": "string",
                        "example": "Ukrainian poet, artist, and thinker. Born into a family of serfs, he gained his freedom through a ransom. His work became the foundation of modern Ukrainian literature.",
                        "nullable": true
                    },
                    "image_url": {
                        "description": "Image url of the author.",
                        "type": "string",
                        "format": "uri",
                        "example": "http://loclhost:8080/storage/authors/author1.png",
                        "nullable": true
                    },
                    "books": {
                        "description": "Books of the author. Returns an empty array if not loaded. Loaded automatically in \"show\" or via \"?include=books\" in \"index\".",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BookResponse"
                        },
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}